requirements.txt
python-telegram-bot
openai
nest_asyncio
python-dotenv
flask
